requirements.txt
pandas
openpyxl
PyMuPDF
opencv-python
pytesseract

python-pptx
python-docx
python-dotenv
pandas
openpyxl
PyMuPDF
opencv-python
pytesseract

python-pptx
python-docx
python-dotenv
spacy
google-generativeai
